Gameplay loop

+++ scenes/TileAdder.gd
@@ -13,9 +13,9 @@ onready var tween = $Tween
 func _ready():
 	pass
 
-func _input(event):
-	if event is InputEventMouseButton and event.button_index == BUTTON_LEFT:
-		var mouse_pos = get_viewport().get_mouse_position();
+func _unhandled_input(event):
+	if event is InputEventMouseButton and Input.is_action_pressed("mouse_left"):
+		var mouse_pos = get_viewport().get_mouse_position()
 		var pos = plane.intersects_ray(
 			camera.project_ray_origin(mouse_pos),
 			camera.project_ray_normal(mouse_pos)
@@ -25,7 +25,6 @@ func _input(event):
 			pos.z = round(pos.z)
 			if tiles.find(Vector2(pos.x, pos.z)) == -1 and current_tile != null:
 				current_tile.translation = pos
-				add_child(current_tile)
 				tiles.append(Vector2(pos.x, pos.z))
 				current_tile = null
 
@@ -35,14 +34,22 @@ func _process(_delta):
 	if (pos3d != null):
 		pos3d.x = round(pos3d.x)
 		pos3d.z = round(pos3d.z)
-		if last_pos != pos3d and current_tile:
+		if last_pos != pos3d and current_tile and tiles.find(Vector2(pos3d.x, pos3d.z)) == -1:
 			# animate movement
 			if tween.is_active():
 				tween.stop_all()
 				tween.remove_all()
-				last_pos = translation
+				last_pos = current_tile.translation
 			tween.interpolate_property(current_tile, "translation",
 				last_pos, pos3d, transition_speed,
 				Tween.TRANS_LINEAR, Tween.EASE_IN_OUT)
 			tween.start()
 			last_pos = pos3d
+
+
+func _on_roll(tile: Spatial):
+	if current_tile != null:
+		print("Rolled before we placed previous tile")
+		return
+	current_tile = tile
+	add_child(current_tile)
